clean up themes
