+++ frontend/app/layout.tsx
@@ -61,7 +61,7 @@ export default async function RootLayout({ children }: ChildrenProps & AppProps)
 
               <div className="flex flex-col items-center min-h-screen ">
                 <AwesomeParticles />
-                <div className="flex flex-col items-center min-h-screen backdrop-blur shadow-2xl shadow-white w-5/6 ">
+                <div className="flex flex-col items-center min-h-screen backdrop-blur shadow-2xl dark:shadow-white w-5/6 ">
                   <NavigationBar />
 
                   <main className="flex flex-col min-h-screen items-center w-full">
