--- frontend/lib/tailwindcss/themes/common.ts
@@ -1,19 +1,0 @@
-import { ConfigTheme } from "~/components/nextui/client-components";
-
-const CommonTheme: ConfigTheme = {
-  layout: {
-    disabledOpacity: "0.3",
-    radius: {
-      small: "4px",
-      medium: "6px",
-      large: "8px",
-    },
-    borderWidth: {
-      small: "1px",
-      medium: "2px",
-      large: "3px",
-    },
-  },
-};
-
-export { CommonTheme };
