+++ frontend/tailwind.config.ts
@@ -1,9 +1,23 @@
 import type { Config } from "tailwindcss";
 import { withUt } from "uploadthing/tw";
 
-import { nextui } from "@nextui-org/react";
-import DarkPurple from "./lib/tailwindcss/themes/dark-purple";
-import LightPurple from "./lib/tailwindcss/themes/light-purple";
+import { ConfigTheme, nextui } from "@nextui-org/react";
+
+const common: ConfigTheme = {
+  layout: {
+    disabledOpacity: "0.3",
+    radius: {
+      small: "4px",
+      medium: "6px",
+      large: "8px",
+    },
+    borderWidth: {
+      small: "1px",
+      medium: "2px",
+      large: "3px",
+    },
+  },
+};
 
 const config: Config = {
   darkMode: ["class"],
@@ -14,11 +28,58 @@ const config: Config = {
     "./src/**/*.{ts,tsx}",
     "./node_modules/@nextui-org/theme/dist/**/*.{js,ts,jsx,tsx}",
   ],
+  theme: {
+    extend: {},
+  },
   plugins: [
     nextui({
       themes: {
-        "purple-dark": DarkPurple,
-        "light-purple": LightPurple,
+        light: {
+          ...common,
+          extend: "light",
+          colors: {
+            // background: "#0D001A",
+            // foreground: "#ffffff",
+            primary: {
+              50: "#3B096C",
+              100: "#520F83",
+              200: "#7318A2",
+              300: "#9823C2",
+              400: "#c031e2",
+              500: "#DD62ED",
+              600: "#F182F6",
+              700: "#FCADF9",
+              800: "#FDD5F9",
+              900: "#FEECFE",
+              DEFAULT: "#DD62ED",
+              foreground: "#ffffff",
+            },
+            focus: "#F182F6",
+          },
+        },
+        dark: {
+          ...common,
+          extend: "dark",
+          colors: {
+            background: "#0D001A",
+            foreground: "#ffffff",
+            primary: {
+              50: "#3B096C",
+              100: "#520F83",
+              200: "#7318A2",
+              300: "#9823C2",
+              400: "#c031e2",
+              500: "#DD62ED",
+              600: "#F182F6",
+              700: "#FCADF9",
+              800: "#FDD5F9",
+              900: "#FEECFE",
+              DEFAULT: "#DD62ED",
+              foreground: "#ffffff",
+            },
+            focus: "#F182F6",
+          },
+        },
       },
     }),
   ],
